Adding possibility for bot to work with different users id

+++ expenses.py
@@ -4,59 +4,64 @@ import datetime
 
 class Message:
     """Структура распаршенного сообщения о новом расходе"""
-    def __init__(self, category_text, amount, created):
+    def __init__(self, category_text, amount, created, user_id):
         self.category_text = category_text
         self.amount = amount
         self.created = created
+        self.user_id = user_id
 
 
 def add_expense(message_split : list):
     """Добавляет новое сообщение.
     Принимает на вход текст сообщения, пришедшего в бот."""
-    obj_message = _parse_message(message_split)
+    obj_message = _parse_message(message_split) #obj_message == Message object
     try:
         db.insert(obj_message)
     except:
         raise ConnectionRefusedError("Проблема с базой данных")
 
 
-def get_today_statistics() -> str:
+def get_today_statistics(user_id : int) -> str:
     cursor = db.get_cursor()
-    cursor.execute("""SELECT category_name, SUM(amount)
-                      FROM expenses WHERE date(created) = date(now())	
+    cursor.execute(f"""SELECT category_name, SUM(amount)
+                      FROM expenses WHERE date(created) = date(now())
+                      AND user_id = {user_id}	
                       GROUP BY category_name""")
     result = cursor.fetchall()
     res_in_str = "\n".join([f"{tpl_values[0]}: {tpl_values[1]}" for tpl_values in result])
     return res_in_str
 
 
-def get_month_statistics() -> str:
+def get_month_statistics(user_id : int) -> str:
     cursor = db.get_cursor()
-    cursor.execute("""SELECT category_name, SUM(amount)
+    cursor.execute(f"""SELECT category_name, SUM(amount)
                       FROM expenses where EXTRACT(MONTH FROM created) = 
-                      EXTRACT(MONTH from now())
+                      EXTRACT(MONTH from now()) AND user_id = {user_id}
                       GROUP BY category_name""")
     result = cursor.fetchall()
     res_in_str = "\n".join([f"{tpl_values[0]}: {tpl_values[1]}" for tpl_values in result])
     return res_in_str
 
 
-def get_expenses():
+def get_expenses(user_id : int) -> str:
     cursor = db.get_cursor()
-    cursor.execute("""SELECT c.name, SUM(amount) FROM
+    cursor.execute(f"""SELECT c.name, SUM(amount) FROM
                         expenses e INNER JOIN category c
                         ON e.category_name = c.name
+                        WHERE e.user_id = {user_id}
                         GROUP BY c.name""")
     res = cursor.fetchall()
     res_in_str = "\n".join([f"{tpl_values[0]}: {tpl_values[1]}" for tpl_values in res])
     return res_in_str
 
 
-def clear(message : str):
-    category = message
+def clear(message : list):
+    category = message[0]
+    user_id = message[1]
     cursor = db.get_cursor()
     cursor.execute(f"""DELETE FROM expenses
-                    WHERE category_name = '{message}' """)
+                    WHERE category_name = '{category}'
+                    AND user_id = {user_id}""")
 
 
 def _parse_message(parsed_message: list) -> Message:
@@ -64,7 +69,8 @@ def _parse_message(parsed_message: list) -> Message:
     category_text = parsed_message[0]
     amount = parsed_message[1]
     created = _get_now_formatted()
-    return Message( category_text=category_text, amount=amount, created=created)
+    user_id = parsed_message[2]
+    return Message( category_text=category_text, amount=amount, created=created, user_id=user_id)
 
 
 def _get_now_formatted() -> str:
